requirements.txt
discord
json
pytest
python-dotenv
requests
